requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
